fix car listing menu and refactor code
