--- flutter_web_aula/lib/pages/carros_page.dart
@@ -1,21 +1,0 @@
-import 'package:flutter/material.dart';
-
-class CarrosPage extends StatefulWidget {
-  @override
-  _CarrosPageState createState() => _CarrosPageState();
-}
-
-class _CarrosPageState extends State<CarrosPage> {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: _body(),
-    );
-  }
-
-  _body() {
-    return Center(
-      child: Text("Carros"),
-    );
-  }
-}
